backend/requirements.txt
Flask
Flask-Cors
spotipy
nltk
python-dotenv